+++ inc/simulator.hpp
@@ -97,7 +97,6 @@ As for PackModule, default zero if port is not specified.
     int Simulate();
     int Simulate_FirstStep();
     int Simulate_OneStep();
-    int Simulate_FinalStep();
 
     // Whether to store simulation data to memory.
     // If set false, then function "Simulator::Plot()" won't be executed.
@@ -157,19 +156,6 @@ private:
     // Pointers to every mat modules which belongs to this simulator.
     // Mat modules will be broke up during initialization.
     std::vector<PMatModule> _matmodules;
-
-    // 3 kinds of pointers below are pointers to endpoint modules.
-    // The reason why they should exist is they all have some private members or
-    //  member functions that different to others and should be treated distunguishly.
-    // "_integrators" has private member variations "_outvalue"
-    //  which will be called in "Simulate_OneStep()".
-    // "_outputs" has private member variations "_values"
-    //  which will be called in "Plot()".
-    // "_unitdelays" has private member functions "Output_Update()"
-    //  which will be called in "Simulate_OneStep()".
-    std::vector<PUIntegrator> _integrators;
-    std::vector<PUOutput> _outputs;
-    std::vector<PUUnitDelay> _unitdelays;
 
     // IDs of every Endpoint modules according to the their updating orders.
     // First ID of every vector is an Endpoint module.